LayoutTests/js/dom/modules/module-scope-set-interval-expected.txt

Test relationship between module scope and setInterval

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
PASS visibleAtATime is "Module Scope"
PASS visibleAtATime is "Module Scope"
PASS visibleAtATime is "Global Scope"
PASS successfullyParsed is true

TEST COMPLETE
